Relax dependencies
Allow graphene-django-extras to install newer versions of the listed
dependencies. These were previously pinned due to graphene-django's
incompatibility with django-filter 2.0; this is no longer an issue, so
the dependency pinning should no longer be required.

Additionally, provide a `name` parameter to the graphql URL in the test
project to allow Django's `reverse` function to find it correctly.

Note: One test currently fails, but it fails consistently before and
after these changes.